requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-ollama
faiss-cpu
sentence-transformers
pandas
python-dotenv
serpapi
ollama
streamlit
scikit-learn
google-search-results
requests
